# .specify/specs/rag-chatbot/contracts/api.yaml
openapi: 3.0.0
info:
  title: RAG Chatbot API
  description: API for Physical AI & Humanoid Robotics Textbook Chatbot
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
    description: Local development
  - url: https://your-api-url.com/api
    description: Production
paths:
...
  /chat/query:
    post:
      summary: General book content query
      tags:
        - Chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    QueryRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: User's question
          maxLength: 1000
        user_id:
          type: string
          description: User ID for session management (required for conversation history)
        session_id:
          type: string
          nullable: true
          description: Session ID (optional, will be created if not provided)
        max_chunks:
          type: integer
          default: 5
          description: Maximum number of chunks to retrieve
...
    QueryResponse:
      type: object
      properties:
        answer:
          type: string
          description: Generated answer from agent
        session_id:
          type: string
          description: Session ID for maintaining conversation history
        sources:
          type: array
          items:
            type: object
            properties:
              document_path:
                type: string
              chunk_index:
                type: integer
              relevance_score:
                type: number
        response_time_ms:
          type: integer
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string